Finish command, send tps on Bukkit/Sponge and update data format.

+++ common/src/main/java/com/azuriom/azlink/common/data/WorldData.java
@@ -2,17 +2,17 @@ package com.azuriom.azlink.common.data;
 
 public class WorldData {
 
-    private final int tps;
+    private final double tps;
     private final int chunks;
     private final int entities;
 
-    public WorldData(int tps, int chunks, int entities) {
+    public WorldData(double tps, int chunks, int entities) {
         this.tps = tps;
         this.chunks = chunks;
         this.entities = entities;
     }
 
-    public int getTps() {
+    public double getTps() {
         return tps;
     }
